runtime: split baremetal memory setup
